+++ parquet/src/main/java/org/apache/iceberg/parquet/ParquetUtil.java
@@ -43,7 +43,6 @@ import org.apache.iceberg.types.Type;
 import org.apache.iceberg.types.Types;
 import org.apache.iceberg.util.BinaryUtil;
 import org.apache.iceberg.util.UnicodeUtil;
-import org.apache.parquet.column.ColumnDescriptor;
 import org.apache.parquet.column.Encoding;
 import org.apache.parquet.column.EncodingStats;
 import org.apache.parquet.column.statistics.Statistics;
@@ -53,8 +52,6 @@ import org.apache.parquet.hadoop.metadata.ColumnChunkMetaData;
 import org.apache.parquet.hadoop.metadata.ColumnPath;
 import org.apache.parquet.hadoop.metadata.ParquetMetadata;
 import org.apache.parquet.schema.MessageType;
-import org.apache.parquet.schema.OriginalType;
-import org.apache.parquet.schema.PrimitiveType;
 
 public class ParquetUtil {
   // not meant to be instantiated
@@ -229,97 +226,6 @@ public class ParquetUtil {
           Conversions.toByteBuffer(schema.findType(entry.getKey()), entry.getValue().value()));
     }
     return bufferMap;
-  }
-
-  public static boolean isFixedLengthDecimal(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    return primitive.getOriginalType() != null &&
-        primitive.getOriginalType() == OriginalType.DECIMAL &&
-        (primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.FIXED_LEN_BYTE_ARRAY ||
-            primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.BINARY);
-  }
-
-  public static boolean isIntLongBackedDecimal(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    return primitive.getOriginalType() != null &&
-        primitive.getOriginalType() == OriginalType.DECIMAL &&
-        (primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.INT64 ||
-            primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.INT32);
-  }
-
-  public static boolean isVarWidthType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType != null &&
-        originalType != OriginalType.DECIMAL &&
-        (originalType == OriginalType.ENUM ||
-            originalType == OriginalType.JSON ||
-            originalType == OriginalType.UTF8 ||
-            originalType == OriginalType.BSON)) {
-      return true;
-    }
-    if (originalType == null && primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.BINARY) {
-      return true;
-    }
-    return false;
-  }
-
-  public static boolean isBooleanType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    return originalType == null && primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.BOOLEAN;
-  }
-
-  public static boolean isFixedWidthBinary(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType == null &&
-        primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.FIXED_LEN_BYTE_ARRAY) {
-      return true;
-    }
-    return false;
-  }
-
-  public static boolean isIntType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType != null && (originalType == OriginalType.INT_8 || originalType == OriginalType.INT_16 ||
-        originalType == OriginalType.INT_32 || originalType == OriginalType.DATE)) {
-      return true;
-    } else if (primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.INT32) {
-      return true;
-    }
-    return false;
-  }
-
-  public static boolean isLongType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType != null && (originalType == OriginalType.INT_64 || originalType == OriginalType.TIMESTAMP_MILLIS ||
-        originalType == OriginalType.TIMESTAMP_MICROS)) {
-      return true;
-    } else if (primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.INT64) {
-      return true;
-    }
-    return false;
-  }
-
-  public static boolean isDoubleType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType == null && primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.DOUBLE) {
-      return true;
-    }
-    return false;
-  }
-
-  public static boolean isFloatType(ColumnDescriptor desc) {
-    PrimitiveType primitive = desc.getPrimitiveType();
-    OriginalType originalType = primitive.getOriginalType();
-    if (originalType == null && primitive.getPrimitiveTypeName() == PrimitiveType.PrimitiveTypeName.FLOAT) {
-      return true;
-    }
-    return false;
   }
 
   @SuppressWarnings("deprecation")
